css/animations/automated/animation-direction_alternate-reverse.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-direction, alternate-reverse</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="http://dev.w3.org/csswg/css3-animations/#the-animation-direction-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="alternate-reverse is a valid animation-direction">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         0% {
            left: 0px;
         }
         100% {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
         animation-delay: 0.4s;
         animation-duration: 0.8s;
         animation-iteration-count: 2;
         animation-fill-mode: both;
         animation-direction: alternate-reverse;
      }
      .test {
         animation-name: slide;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 3000);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            passed = passed && (getComputedStyle(div, null).left == '100px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
         var passed = (getComputedStyle(div, null).left == '100px');
      }
   </script>
   <script src="../resources/add_prefixes.js"></script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should slide from right to left and then back
   right.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
